Move util.go to more appropriately named files
Move util.go to more appropriately named files
